Se incorpora archivo para guardar progreso en el juego

+++ Parcial 2/Practica 1 unidad 3/Assets/scripts/contadorDeEnemigos.cs
@@ -1,15 +1,26 @@
 using System.Collections;
 using System.Collections.Generic;
-using UnityEngine;
-using TMPro;
 using System.IO;
+using TMPro;
+using UnityEditor.VersionControl;
+using UnityEngine;
 
 public class contadorDeEnemigos : MonoBehaviour
 {
+    // variables de inferfas grafica
+    public TextMeshProUGUI texto;
+    public TextMeshProUGUI texto2;
+    public TextMeshProUGUI mensaje;
+    public TMP_InputField nombre;
+    // variables comunes
     public static int contEnemigos = 0;
     static StreamWriter archivo = null;
     static StreamReader leer = null;
     public static string dato = "0";
+    public GameObject panel;
+     // bool est = false;
+
+
     // Start is called before the first frame update
     public static void EnemigosEliminados()
     {
@@ -18,11 +29,28 @@ public class contadorDeEnemigos : MonoBehaviour
         //CrearAchivo();
         
     }
-    public TextMeshProUGUI texto;
+   
 
     private void Update()
     {
-        texto.text = "Enemigos Eliminados: " + contEnemigos.ToString();
+        if (Input.GetKeyDown(KeyCode.G))
+        {
+            print("Guardar Datos");
+            //panel.SetActive(!est);
+            panel.SetActive(!panel.activeSelf); // negamos el estado original de panel
+            Time.timeScale = 0f; // pausa el juego
+            print("pumm");
+            texto2.text = "Datos Guardados: " + contEnemigos.ToString();
+            if (panel.activeSelf)
+            {
+                mensaje.text = "Mensaje..."; // linea experimental
+            }
+        }
+        else if (panel.activeSelf == false)
+        {
+            Time.timeScale = 1f; // reaunida el juego
+        }
+            texto.text = "Enemigos Eliminados: " + contEnemigos.ToString();
 
     }
     private void Start()
@@ -37,28 +65,34 @@ public class contadorDeEnemigos : MonoBehaviour
         if (File.Exists("Puntaje.txt"))
         {
             leer = File.OpenText("Puntaje.txt");
-            //do
-            //{
-                //if (leer.ReadLine() != null)
-                //{
-                    print(dato);
-                    dato = leer.ReadLine();
-                //}
-            //}
-            //while (leer != null);
+            //print("Leemos el archivo");
+            print(dato);
+            dato = leer.ReadLine();
+            print("Extraemos dato");
+            print(dato);
             leer.Close();
-            contEnemigos = int.Parse(dato);
+            string[] d = dato.Split(" , ");
+      
+            contEnemigos = int.Parse(d[0]);
+            nombre.text = d[1];
         }
     }
 
-    public static void CrearAchivo()
+    public  void CrearAchivo()
     {
+        string n = nombre.text;
+        if (n.Length == 0)
+        {
+            mensaje.text = "No tiene nombre";
+            return;
+        }
+        mensaje.text = "Guardado";
         if (File.Exists("Puntaje.txt"))
         {
             File.Delete("Puntaje.txt");
         }
         archivo = File.AppendText("Puntaje.txt");
-        archivo.WriteLine(contEnemigos);
+        archivo.WriteLine(contEnemigos + " , " + nombre.text);
         archivo.Close();
     }
 }
